+++ web-client/src/applicationContexts/dev.js
@@ -5,7 +5,7 @@ const API_URL = 'http://localhost:3000/v1';
 /**
  * Context for the dev environment
  */
-const dev = {
+const applicationContext = {
   getBaseUrl: () => {
     return API_URL;
   },
@@ -14,4 +14,4 @@ const dev = {
   },
 };
 
-export default dev;
+export default applicationContext;
