+++ web-client/src/applicationContexts/prod.js
@@ -5,7 +5,7 @@ const API_URL = process.env.API_URL || 'http://localhost:8080';
 /**
  * Context for the prod environment
  */
-const prod = {
+const applicationContext = {
   getBaseUrl: () => {
     return API_URL;
   },
@@ -14,4 +14,4 @@ const prod = {
   },
 };
 
-export default prod;
+export default applicationContext;
